# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backend server for Digital Achievement Ledger.",
        "title": "Digital Achievement Ledger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/certificates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of certificates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certificates"
                ],
                "summary": "Get all certificates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by student ID",
                        "name": "student_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by certificate status (active, revoked)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Issue a new certificate for a student with selected achievements",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certificates"
                ],
                "summary": "Create a certificate",
                "parameters": [
                    {
                        "description": "Certificate creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_model.CreateCertificateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/certificates/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a certificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certificates"
                ],
                "summary": "Get certificate by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/certificates/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate and download the PDF for a specific certificate",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "certificates"
                ],
                "summary": "Download certificate PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certificate PDF file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
        "/certificates/{id}/revoke": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke an issued certificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "certificates"
                ],
                "summary": "Revoke a certificate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        },
...
        "/verify/{token}": {
            "get": {
                "description": "Public verify endpoint for a certificate token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Verify a certificate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Verification Token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_ahmadqo_digital-achievement-ledger_internal_model.CreateCertificateRequest": {
            "type": "object",
            "properties": {
                "achievement_ids": {
                    "description": "prestasi yang dimasukkan ke surat",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "valid_until": {
                    "description": "format: YYYY-MM-DD, opsional",
                    "type": "string"
                }
            }
        },
...
        "github_com_ahmadqo_digital-achievement-ledger_internal_response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_ahmadqo_digital-achievement-ledger_internal_response.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_ahmadqo_digital-achievement-ledger_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_ahmadqo_digital-achievement-ledger_internal_response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}